Extract common test service logic into base class
