+++ src/OmniSharp.DotNetTest/Services/GetDotNetTestStartInfoService.cs
@@ -1,6 +1,4 @@
 using System.Composition;
-using System.IO;
-using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.Extensions.Logging;
 using OmniSharp.DotNetTest.Models;
@@ -10,29 +8,17 @@ using OmniSharp.Services;
 namespace OmniSharp.DotNetTest.Services
 {
     [OmniSharpHandler(OmnisharpEndpoints.V2.GetDotNetTestStartInfo, LanguageNames.CSharp)]
-    public class GetDotNetTestStartInfoService : RequestHandler<GetDotNetTestStartInfoRequest, GetDotNetTestStartInfoResponse>
+    public class GetDotNetTestStartInfoService : BaseTestService<GetDotNetTestStartInfoRequest, GetDotNetTestStartInfoResponse>
     {
-        private readonly OmniSharpWorkspace _workspace;
-        private readonly DotNetCliService _dotNetCli;
-        private readonly ILoggerFactory _loggerFactory;
-
         [ImportingConstructor]
         public GetDotNetTestStartInfoService(OmniSharpWorkspace workspace, DotNetCliService dotNetCli, ILoggerFactory loggerFactory)
+            : base(workspace, dotNetCli, loggerFactory)
         {
-            _workspace = workspace;
-            _dotNetCli = dotNetCli;
-            _loggerFactory = loggerFactory;
         }
 
-        public Task<GetDotNetTestStartInfoResponse> Handle(GetDotNetTestStartInfoRequest request)
+        protected override GetDotNetTestStartInfoResponse HandleRequest(GetDotNetTestStartInfoRequest request, TestManager testManager)
         {
-            var document = _workspace.GetDocument(request.FileName);
-
-            using (var dtm = TestManager.Start(document.Project, _dotNetCli, _loggerFactory))
-            {
-                var response = dtm.GetTestStartInfo(request.MethodName, request.TestFrameworkName);
-                return Task.FromResult(response);
-            }
+            return testManager.GetTestStartInfo(request.MethodName, request.TestFrameworkName);
         }
     }
 }
